refactor: export credential loading mechanism.
