+++ cmd/terramate/cli/tmcloud/_test_mock.tf
@@ -1,0 +1,14 @@
+// TERRAMATE: GENERATED AUTOMATICALLY DO NOT EDIT
+
+resource "local_file" "tmcloud" {
+  content = <<-EOT
+package tmcloud // import "github.com/terramate-io/terramate/cmd/terramate/cli/tmcloud"
+
+Package tmcloud provides helper functions for interacting with the Terramate
+Cloud API from the CLI.
+
+func BaseURL() string
+EOT
+
+  filename = "${path.module}/mock-tmcloud.ignore"
+}
