+++ cmd/terramate/cli/tmcloud/auth/doc.go
@@ -1,0 +1,5 @@
+// Copyright 2025 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+// Package auth provides the helper functions for loading the Terramate Cloud credentials.
+package auth
