+++ cmd/terramate/cli/tmcloud/doc.go
@@ -1,0 +1,5 @@
+// Copyright 2025 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+// Package tmcloud provides helper functions for interacting with the Terramate Cloud API from the CLI.
+package tmcloud
