+++ cmd/terramate/cli/tmcloud/stack.tm.hcl
@@ -1,0 +1,9 @@
+// Copyright 2025 Terramate GmbH
+// SPDX-License-Identifier: MPL-2.0
+
+stack {
+  name        = "package tmcloud // import \"github.com/terramate-io/terramate/cmd/terramate/cli/tmcloud\""
+  description = "package tmcloud // import \"github.com/terramate-io/terramate/cmd/terramate/cli/tmcloud\"\n\nPackage tmcloud provides helper functions for interacting with the Terramate\nCloud API from the CLI.\n\nfunc BaseURL() string"
+  tags        = ["cli", "cmd", "golang", "terramate", "tmcloud"]
+  id          = "099b4271-bf70-4df6-ab29-fe83077b7864"
+}
